requirements.txt
# Core
python-dotenv>=1.0
click>=8.1
jinja2>=3.1
pyyaml>=6.0
duckdb>=0.9
plotly>=5.18
rich>=13.0

# LLM APIs
openai>=1.12
anthropic>=0.18

# Embeddings for semantic similarity
numpy>=1.24
scikit-learn>=1.3

# Token counting
tiktoken>=0.5

# Statistical testing
scipy>=1.11

# Dev
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.23
ruff>=0.1
pre-commit>=3.0
